database/core.py
from tortoise import fields, models

class CoreModel(models.Model):
    id = fields.IntField(pk=True)
    created_at = fields.DatetimeField(auto_now_add=True)
    updated_at = fields.DatetimeField(auto_now=True)

    class Meta:
        abstract = True   # 关键：声明为抽象类，不会单独生成表