pyproject.toml
[project]
name = "15-dbsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.12.0",
    "fire>=0.7.1",
    "msgpack>=1.1.2",
    "sqlite-utils>=3.39",
    "tortoise-orm>=1.1.2",
    "tzdata>=2025.3",
    "wakepy>=1.0.0",
    "zmq>=0.0.0",
]
